% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_multicore.R
\name{batchtools_multicore}
\alias{batchtools_multicore}
\title{A batchtools backend that resolves futures in parallel via forked background R processes}
\usage{
batchtools_multicore(..., workers = availableCores(constraints = "multicore"))
}
\arguments{
\item{workers}{The number of multicore processes to be
available for concurrent batchtools multicore futures.}

\item{\ldots}{Not used.}
}
\description{
A batchtools backend that resolves futures in parallel via forked background R processes
}
\details{
Batchtools multicore futures use \pkg{batchtools} cluster functions
created by \code{\link[batchtools:makeClusterFunctionsMulticore]{batchtools::makeClusterFunctionsMulticore()}} with
\code{ncpus = workers}.

An alternative to the batchtools multicore backend is to use
\code{plan(future::multicore)}.
}
\examples{
plan(batchtools_multicore, workers = 2)

message("Main process ID: ", Sys.getpid())

f <- future(Sys.getpid())
pid <- value(f)
message("Worker process ID: ", pid)

}
