% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_ssh.R
\name{batchtools_ssh}
\alias{batchtools_ssh}
\title{A batchtools backend that resolves futures in parallel via background R sessions over SSH}
\usage{
batchtools_ssh(..., workers = availableWorkers())
}
\arguments{
\item{workers}{The number of multicore processes to be
available for concurrent batchtools multicore futures.}
}
\description{
A batchtools backend that resolves futures in parallel via background R sessions over SSH
}
\details{
The \code{batchtools_ssh} backend uses the batchtools backend set
up by \code{\link[batchtools:makeClusterFunctionsSSH]{batchtools::makeClusterFunctionsSSH()}}, which requires
system commands \code{ssh} and \code{ps} as available on Linux and macOS.

An alternative to \code{batchtools_ssh} is to use
\link[future:cluster]{cluster} futures of the \pkg{future}
package with a single local background session, i.e.
\code{plan(cluster, workers = c("localhost"))}.
}
\keyword{internal}
