% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_custom.R
\name{BatchtoolsCustomFutureBackend}
\alias{BatchtoolsCustomFutureBackend}
\alias{batchtools_custom}
\title{Batchtools futures for custom batchtools configuration}
\usage{
BatchtoolsCustomFutureBackend(...)
}
\arguments{
\item{\ldots}{Additional arguments passed to \code{\link[=BatchtoolsFutureBackend]{BatchtoolsFutureBackend()}}.}
}
\value{
An object of class \code{BatchtoolsFuture}.
}
\description{
Batchtools futures for custom batchtools configuration
}
\examples{
options(error = function(...) {
  print(traceback())
})

cf <- batchtools::makeClusterFunctionsInteractive(external = TRUE)
print(cf)
str(cf)
plan(batchtools_custom, cluster.functions = cf)
print(plan())
print(nbrOfWorkers())

## Create explicit future
f <- future({
  cat("PID:", Sys.getpid(), "\n")
  42L
})
print(f)
v <- value(f)
print(v)

options(error = NULL)


## Create explicit future
f <- future({
  cat("PID:", Sys.getpid(), "\n")
  42L
})
print(f)
v <- value(f)
print(v)



## Create explicit future
f <- future({
  cat("PID:", Sys.getpid(), "\n")
  42L
})
v <- value(f)
print(v)
}
