% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchtoolsFutureBackend-class.R
\name{BatchtoolsFutureBackend}
\alias{BatchtoolsFutureBackend}
\alias{BatchtoolsUniprocessFutureBackend}
\alias{BatchtoolsMultiprocessFutureBackend}
\title{A batchtools future is a future whose value will be resolved via batchtools}
\usage{
BatchtoolsFutureBackend(
  workers = NULL,
  resources = list(),
  finalize = getOption("future.finalize", TRUE),
  conf.file = findConfFile(),
  cluster.functions = NULL,
  registry = list(),
  interrupts = TRUE,
  ...
)
}
\arguments{
\item{workers}{(optional) The maximum number of workers the batchtools
backend may use at any time.   Interactive and "local" backends can only
process one future at the time (\code{workers = 1L}), whereas HPC backends,
where futures are resolved via separate jobs on a scheduler, can have
multiple workers.  In the latter, the default is \code{workers = NULL}, which
will resolve to
\code{getOption("\link{future.batchtools.workers}")}.
If neither are specified, then the default is \code{100}.}

\item{resources}{(optional) A named list passed to the \pkg{batchtools}
template (available as variable \code{resources}).  See Section 'Resources'
in \code{\link[batchtools:submitJobs]{batchtools::submitJobs()}} more details.}

\item{finalize}{If TRUE, any underlying registries are
deleted when this object is garbage collected, otherwise not.}

\item{conf.file}{(optional) A batchtools configuration file.}

\item{cluster.functions}{(optional) A batchtools
\link[batchtools:makeClusterFunctions]{ClusterFunctions} object.}

\item{registry}{(optional) A named list of settings to control the setup
of the batchtools registry.}

\item{interrupts}{If FALSE, attempts to interrupt futures will not take
place on this backend, even if the backend supports it. This is useful
when, for instance, it takes a long time to interrupt a future.}

\item{\ldots}{Not used.}
}
\value{
A \link[future:FutureBackend]{future::FutureBackend} object of class BatchtoolsFutureBackend
}
\description{
A batchtools future is a future whose value will be resolved via batchtools
}
\keyword{internal}
