% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchtoolsFutureError.R
\name{BatchtoolsFutureError}
\alias{BatchtoolsFutureError}
\title{FutureError class for errors related to BatchtoolsFutures}
\usage{
BatchtoolsFutureError(...)
}
\arguments{
\item{\ldots}{Arguments passed to \link[future:FutureError]{FutureError}.}
}
\description{
FutureError class for errors related to BatchtoolsFutures
}
\keyword{internal}
