% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_ssh.R
\name{BatchtoolsSSHFutureBackend}
\alias{BatchtoolsSSHFutureBackend}
\title{batchtools SSH futures}
\usage{
BatchtoolsSSHFutureBackend(workers = availableWorkers(), ...)
}
\arguments{
\item{workers}{The number of SSH processes to be
available for concurrent batchtools SSH futures.}

\item{\ldots}{Additional arguments passed
to \code{\link[=BatchtoolsFutureBackend]{BatchtoolsFutureBackend()}}.}
}
\value{
An object of class \code{BatchtoolsMulticoreFuture}.
}
\description{
A batchtools SSH future is an asynchronous multiprocess
future that will be evaluated in a background R session.\cr
\cr
\emph{We highly recommend using \link[future:multisession]{future::multisession}
(sic!) futures of the \pkg{future} package instead of
SSH batchtools futures.}
}
\details{
batchtools SSH futures rely on the batchtools backend set
up by \code{\link[batchtools:makeClusterFunctionsSSH]{batchtools::makeClusterFunctionsSSH()}}.
The batchtools SSH backend only works on operating systems
supporting the \code{ssh} and \code{ps} command-line tool, e.g. Linux and macOS.
}
\keyword{internal}
