% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchtoolsFuture-class.R
\name{loggedOutput}
\alias{loggedOutput}
\alias{loggedError}
\alias{loggedError.BatchtoolsFuture}
\title{Logged output of batchtools future}
\usage{
loggedOutput(...)

\method{loggedError}{BatchtoolsFuture}(future, ...)
}
\arguments{
\item{future}{The future.}

\item{\ldots}{Not used.}
}
\value{
A character vector or a logical scalar.
}
\description{
Logged output of batchtools future
}
\keyword{internal}
