% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbrOfWorkers.R
\name{nbrOfWorkers.batchtools}
\alias{nbrOfWorkers.batchtools}
\alias{nbrOfWorkers.batchtools_local}
\alias{nbrOfWorkers.batchtools_interactive}
\alias{nbrOfWorkers.batchtools_multicore}
\alias{nbrOfWorkers.batchtools_custom}
\alias{nbrOfWorkers.batchtools_lsf}
\alias{nbrOfWorkers.batchtools_openlava}
\alias{nbrOfWorkers.batchtools_sge}
\alias{nbrOfWorkers.batchtools_slurm}
\alias{nbrOfWorkers.batchtools_torque}
\title{Gets the number of batchtools workers}
\usage{
\method{nbrOfWorkers}{batchtools}(evaluator)
}
\arguments{
\item{evaluator}{A future evaluator function.
If NULL (default), the current evaluator as returned
by \code{\link{plan}()} is used.}
}
\value{
A number in [1,Inf].
}
\description{
Tries to infer the total number of batchtools workers.  This is
done using various ad hoc procedures based on code inspection
of batchtools itself.
}
\keyword{internal}
