% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchtoolsFuture-class.R
\name{print.BatchtoolsFuture}
\alias{print.BatchtoolsFuture}
\title{Prints a batchtools future}
\usage{
\method{print}{BatchtoolsFuture}(x, ...)
}
\arguments{
\item{x}{An BatchtoolsFuture object}

\item{\ldots}{Not used.}
}
\description{
Prints a batchtools future
}
\keyword{internal}
