% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchtoolsFuture-class.R
\name{status.BatchtoolsFuture}
\alias{status.BatchtoolsFuture}
\alias{status}
\alias{finished}
\alias{result}
\alias{loggedError}
\alias{loggedOutput}
\title{Status of batchtools future}
\usage{
\method{status}{BatchtoolsFuture}(future, ...)
}
\arguments{
\item{future}{The future.}

\item{\ldots}{Not used.}
}
\value{
A character vector or a logical scalar.
}
\description{
Status of batchtools future
}
\keyword{internal}
