% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_custom.R
\name{batchtools_custom}
\alias{batchtools_custom}
\title{batchtools conf futures}
\usage{
batchtools_custom(expr, envir = parent.frame(), substitute = TRUE,
  globals = TRUE, label = "batchtools", conf = NULL, pathname = NULL,
  workers = NULL, resources = list(), ...)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which global environment
should be located.}

\item{substitute}{Controls whether \code{expr} should be
\code{substitute()}:d or not.}

\item{globals}{(optional) a logical, a character vector, a named list, or a \link[globals]{Globals} object.  If TRUE, globals are identified by code inspection based on \code{expr} and \code{tweak} searching from environment \code{envir}.  If FALSE, no globals are used.  If a character vector, then globals are identified by lookup based their names \code{globals} searching from environment \code{envir}.  If a named list or a Globals object, the globals are used as is.}

\item{label}{(optional) Label of the future (where applicable, becomes the job name for most job schedulers).}

\item{conf}{A batchtools configuration environment.}

\item{pathname}{(alternative) Pathname to one or more batchtools
configuration files to be loaded in order.  If NULL, then the
\pkg{batchtools} package will search for such configuration files.}

\item{workers}{(optional) Additional specification for the backend
workers.  If NULL, the default is used.}

\item{resources}{A named list passed to the batchtools template (available as variable \code{resources}).}

\item{\ldots}{Additional arguments passed to \code{\link{BatchtoolsFuture}()}.}
}
\value{
An object of class \code{BatchtoolsFuture}.
}
\description{
A conf batchtools future sources one or more batchtools configuration
files (R source scripts) to define the batchtools configuration
environment, e.g. \file{.batchtools.R}.
}
\details{
If \code{conf} is NULL (default), then the batchtools configuration will
be created from a set of batchtools configuration files (R script files)
as given by argument \code{pathname}.  If none are specified (default),
then \pkg{batchtools} is designed to use (in order) all of following
configuration files (if they exist):
\itemize{
 \item \code{system("etc", "batchtools_global_config.R", package="batchtools")}
 \item \code{~/.batchtools.R} (in user's home directory)
 \item \code{.batchtools.R} (in the current directory)
}
}
