% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_lsf.R
\name{batchtools_lsf}
\alias{batchtools_lsf}
\title{A batchtools LSF backend resolves futures in parallel via a Load Sharing Facility (LSF) job scheduler}
\usage{
batchtools_lsf(
  ...,
  template = "lsf",
  scheduler.latency = 1,
  fs.latency = 65,
  resources = list()
)
}
\arguments{
\item{template}{(optional) Name of job-script template to be searched
for by \code{\link[batchtools:findTemplateFile]{batchtools::findTemplateFile()}}. If not found, it defaults to
the \code{templates/lsf.tmpl} part of this package (see below).}

\item{scheduler.latency}{[\code{numeric(1)}]\cr
Time to sleep after important interactions with the scheduler to ensure a sane state.
Currently only triggered after calling \code{\link[batchtools]{submitJobs}}.}

\item{fs.latency}{[\code{numeric(1)}]\cr
Expected maximum latency of the file system, in seconds.
Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
access files and directories.
Usually safe to set to \code{0} to disable the heuristic, e.g. if you are working on a local file system.}

\item{resources}{(optional) A named list passed to the \pkg{batchtools}
job-script template as variable \code{resources}.  See Section 'Resources'
in \code{\link[batchtools:submitJobs]{batchtools::submitJobs()}} more details.}

\item{\ldots}{Not used.}
}
\description{
A batchtools LSF backend resolves futures in parallel via a Load Sharing Facility (LSF) job scheduler
}
\details{
Batchtools lsf futures use \pkg{batchtools} cluster functions
created by \code{\link[batchtools:makeClusterFunctionsLSF]{batchtools::makeClusterFunctionsLSF()}}, which requires
that LSF commands \code{bsub}, \code{bjobs}, and \code{bkill} are installed on
the current machine.

The default template script \code{templates/lsf.tmpl} can be found in:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("templates", "lsf.tmpl", package = "future.batchtools")
}\if{html}{\out{</div>}}

and comprise:

\if{html}{\out{<div class="sourceCode bash">}}\preformatted{#!/bin/bash
######################################################################
# A batchtools launch script template for LSF
#
# Author: Henrik Bengtsson 
######################################################################

## Job name:
#BSUB -J <\%= job.name \%>

## Direct streams to logfile:
#BSUB -o <\%= log.file \%>

## Resources needed:
<\% if (length(resources) > 0) \{
  opts <- unlist(resources, use.names = TRUE)
  opts <- sprintf("-\%s=\%s", names(opts), opts)
  writeLines(sprintf("#BSUB \%s", opts))
\} \%>

echo "Batchtools information:"
echo "- job name: '<\%= job.name \%>'"
echo "- job log file: '<\%= log.file \%>'"
echo

echo "Session information:"
echo "- timestamp: $(date)"
echo "- hostname: $(hostname)"
echo "- Rscript path: $(which Rscript)"
echo "- Rscript version: $(Rscript --version)"
echo "- Rscript library paths: $(Rscript -e "cat(shQuote(.libPaths()), sep = ' ')")"
echo

echo "Job submission declarations:"
<\%
  writeLines(sprintf("echo '#BSUB \%s'", opts))
\%>

echo "Job information:"
bjobs -l "$\{LSB_JOBID\}"
echo

## Launch R and evaluate the batchtools R job
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ..."
Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")'
res=$?
echo " - exit code: $\{res\}"
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ... done"

## End-of-job summary
bjobs -l "$\{LSB_JOBID\}"

## Relay the exit code from Rscript
exit "$\{res\}"
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Limit runtime to 3 minutes and memory to 200 MiB per future
plan(future.batchtools::batchtools_lsf, resources = list(W = "00:03:00", M = "200"))

message("Main process ID: ", Sys.getpid())

f <- future(Sys.getpid())
pid <- value(f)
message("Worker process ID: ", pid)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \url{https://en.wikipedia.org/wiki/IBM_Spectrum_LSF}
}
}
