% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_openlava.R
\name{batchtools_openlava}
\alias{batchtools_openlava}
\title{A batchtools openlava backend resolves futures in parallel via a OpenLava job scheduler}
\usage{
batchtools_openlava(..., template = "openlava")
}
\arguments{
\item{template}{(optional) Name of job-script template to be searched
for by \code{\link[batchtools:findTemplateFile]{batchtools::findTemplateFile()}}. If not found, it defaults to
the \code{templates/openlava.tmpl} part of this package (see below).}

\item{\ldots}{Not used.}
}
\description{
A batchtools openlava backend resolves futures in parallel via a OpenLava job scheduler
}
\details{
Batchtools openlava futures use \pkg{batchtools} cluster functions
created by \code{\link[batchtools:makeClusterFunctionsOpenLava]{batchtools::makeClusterFunctionsOpenLava()}}, which requires
that OpenLava commands \code{sbatch}, \code{squeue}, and \code{scancel} are installed on
the current machine.

The default template script \code{templates/openlava.tmpl} can be found in:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("templates", "openlava.tmpl", package = "future.batchtools")
}\if{html}{\out{</div>}}

and comprise:

\if{html}{\out{<div class="sourceCode bash">}}\preformatted{#!/bin/bash
######################################################################
# A batchtools launch script template for OpenLava
#
# Author: Henrik Bengtsson 
######################################################################

## Job name:
#BSUB -J <\%= job.name \%>

## Direct streams to logfile:
#BSUB -o <\%= log.file \%>

## Resources needed:
<\% if (length(resources) > 0) \{
  opts <- unlist(resources, use.names = TRUE)
  opts <- sprintf("-\%s=\%s", names(opts), opts)
  cat(sprintf("#BSUB \%s\\n", opts))
<\% \} \%>

echo "Batchtools job name: '<\%= job.name \%>'"

echo "Session information:"
date
hostname
which Rscript
Rscript --version
Rscript -e ".libPaths()"

## Launch R and evaluate the batchtools R job
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ..."
Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")'
res=$?
echo " - exit code: $\{res\}"
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ... done"

## End-of-job summary
bjobs -l "$\{LSB_JOBID\}"

## Relay the exit code from Rscript
exit "$\{res\}"
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Limit runtime to 3 minutes and memory to 200 MiB per future
plan(batchtools_lsf, resources = list(W = "00:03:00", M = "200"))

message("Main process ID: ", Sys.getpid())

f <- future(Sys.getpid())
pid <- value(f)
message("Worker process ID: ", pid)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \url{https://en.wikipedia.org/wiki/OpenLava}
}
}
