% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_sge.R
\name{batchtools_sge}
\alias{batchtools_sge}
\title{A batchtools SGE backend resolves futures in parallel via a SGE job scheduler}
\usage{
batchtools_sge(
  ...,
  template = "sge",
  scheduler.latency = 1,
  fs.latency = 65,
  resources = list(),
  workers = getOption("future.batchtools.workers", default = 100L)
)
}
\arguments{
\item{template}{(optional) Name of job-script template to be searched
for by \code{\link[batchtools:findTemplateFile]{batchtools::findTemplateFile()}}. If not found, it defaults to
the \code{templates/sge.tmpl} part of this package (see below).}

\item{scheduler.latency}{[\code{numeric(1)}]\cr
Time to sleep after important interactions with the scheduler to ensure a sane state.
Currently only triggered after calling \code{\link[batchtools]{submitJobs}}.}

\item{fs.latency}{[\code{numeric(1)}]\cr
Expected maximum latency of the file system, in seconds.
Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
access files and directories.
Usually safe to set to \code{0} to disable the heuristic, e.g. if you are working on a local file system.}

\item{workers}{The maximum number of workers the batchtools backend may
use at any time, which for HPC schedulers corresponds to the maximum number
of queued jobs. The default is
\code{getOption("\link{future.batchtools.workers}", 100)}.}

\item{\ldots}{Not used.}
}
\description{
A batchtools SGE backend resolves futures in parallel via a SGE job scheduler
}
\details{
Batchtools sge futures use \pkg{batchtools} cluster functions
created by \code{\link[batchtools:makeClusterFunctionsSGE]{batchtools::makeClusterFunctionsSGE()}}, which requires
that SGE commands \code{qsub}, \code{qstat}, and \code{qdel} are installed on
the current machine.

The default template script \code{templates/sge.tmpl} can be found in:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("templates", "sge.tmpl", package = "future.batchtools")
}\if{html}{\out{</div>}}

and comprise:

\if{html}{\out{<div class="sourceCode bash">}}\preformatted{#!/bin/bash
######################################################################
# A batchtools launch script template for SGE
#
# Author: Henrik Bengtsson 
######################################################################
## Shell:
#$ -S /bin/bash         # Run this as a bash script (required)

## Job name:
#$ -N <\%= job.name \%>

## Direct streams to logfile:
#$ -o <\%= log.file \%>

## Merge standard error and output:
#$ -j y

## Tell the queue system to use the current directory
## as the working directory
#$ -cwd

## Use environment variables
#$ -V

## Resources needed:
<\% if (length(resources) > 0) \{
  ## As-is resource specifications
  if ("asis" \%in\% names(resources)) \{
    lines <- resources[["asis"]]
    resources <- resources[names(resources) != "asis"]
  \} else \{
    lines <- character(0)
  \}
     
  ## Remaining resources are assumed to be of type '-l <key>=<value>'
  opts <- unlist(resources, use.names = TRUE)
  opts <- sprintf("-l \%s=\%s", names(opts), opts)
  lines <- sprintf("#$ \%s", c(lines, opts))
  writeLines(lines)
\} \%>

echo "Batchtools information:"
echo "- job name: '<\%= job.name \%>'"
echo "- job log file: '<\%= log.file \%>'"
echo

echo "Session information:"
echo "- timestamp: $(date)"
echo "- hostname: $(hostname)"
echo "- Rscript path: $(which Rscript)"
echo "- Rscript version: $(Rscript --version)"
echo "- Rscript library paths: $(Rscript -e "cat(shQuote(.libPaths()), sep = ' ')")"
echo

echo "Job submission declarations:"
<\%
  writeLines(sprintf("echo '\%s'", lines))
\%>

echo "Job information:"
qstat -j "$\{JOB_ID\}"
echo

## Launch R and evaluate the batchtools R job
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ..."
Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")'
res=$?
echo " - exit code: $\{res\}"
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ... done"

## End-of-job summary
qstat -j "$\{JOB_ID\}"

## Relay the exit code from Rscript
exit "$\{res\}"
}\if{html}{\out{</div>}}

This template and the built-in \code{\link[batchtools:makeClusterFunctionsSGE]{batchtools::makeClusterFunctionsSGE()}}
have been verified to work with SGE 8.1.9 (Son of Grid Engine) running on
a Rocky 8 Linux cluster with a BeeGFS global filesystem.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(future)

# Limit runtime to 3 minutes and memory to 200 MiB per future,
# request two slots and run in the current working directory.
plan(future.batchtools::batchtools_sge, resources = list(
  h_rt = "00:03:00", mem_free = "200M",
  asis = c("-pe smp 2", "-cwd")
))

f <- future(Sys.info())
info <- value(f)
print(info)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \url{https://en.wikipedia.org/wiki/Oracle_Grid_Engine}
}
}
