% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_slurm.R
\name{batchtools_slurm}
\alias{batchtools_slurm}
\title{A batchtools slurm backend resolves futures in parallel via a Slurm job scheduler}
\usage{
batchtools_slurm(...)
}
\arguments{
\item{template}{(optional) Name of job-script template to be searched
for by \code{\link[batchtools:findTemplateFile]{batchtools::findTemplateFile()}}. If not found, it defaults to
the \code{templates/slurm.tmpl} part of this package (see below).}

\item{\ldots}{Not used.}
}
\description{
A batchtools slurm backend resolves futures in parallel via a Slurm job scheduler
}
\details{
Batchtools slurm futures use \pkg{batchtools} cluster functions
created by \code{\link[batchtools:makeClusterFunctionsSlurm]{batchtools::makeClusterFunctionsSlurm()}}, which requires
that Slurm commands \code{sbatch}, \code{squeue}, and \code{scancel} are installed on
the current machine.

The default template script \code{templates/slurm.tmpl} can be found in:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("templates", "slurm.tmpl", package = "future.batchtools")
}\if{html}{\out{</div>}}

and comprise:

\if{html}{\out{<div class="sourceCode bash">}}\preformatted{#!/bin/bash
######################################################################
# A batchtools launch script template for Slurm
#
# Author: Henrik Bengtsson 
######################################################################

<\% Slurm defaults
defaults <- list(
  nodes = 1,          # single-host processing
   time = "00:05:00"  # 5-min runtime
   mem  = "100M"      # 100 MiB memory
)
resources <- c(resources, defaults[setdiff(names(defaults), names(resources))])
opts <- unlist(resources, use.names = TRUE)
opts <- sprintf("--\%s=\%s", names(opts), opts)
opts <- paste(opts, collapse = " ") \%>
\%>

#SBATCH --job-name=<\%= job.name \%>
#SBATCH --output=<\%= log.file \%>
#SBATCH <\%= opts \%>

## Launch R and evaluate the batchtools R job
Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")'
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Limit runtime to 3 minutes and memory to 200 MiB per future
plan(batchtools_slurm, resources = list(time = "00:03:00", mem = "200M"))

message("Main process ID: ", Sys.getpid())

f <- future(Sys.getpid())
pid <- value(f)
message("Worker process ID: ", pid)
\dontshow{\}) # examplesIf}
}
