% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_torque.R
\name{batchtools_torque}
\alias{batchtools_torque}
\title{A batchtools TORQUE backend resolves futures in parallel via a TORQUE/PBS job scheduler}
\usage{
batchtools_torque(
  ...,
  template = "torque",
  scheduler.latency = 1,
  fs.latency = 65
)
}
\arguments{
\item{template}{(optional) Name of job-script template to be searched
for by \code{\link[batchtools:findTemplateFile]{batchtools::findTemplateFile()}}. If not found, it defaults to
the \code{templates/torque.tmpl} part of this package (see below).}

\item{scheduler.latency}{[\code{numeric(1)}]\cr
Time to sleep after important interactions with the scheduler to ensure a sane state.
Currently only triggered after calling \code{\link[batchtools]{submitJobs}}.}

\item{fs.latency}{[\code{numeric(1)}]\cr
Expected maximum latency of the file system, in seconds.
Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to
access files and directories.
Usually safe to set to \code{0} to disable the heuristic, e.g. if you are working on a local file system.}

\item{\ldots}{Not used.}
}
\description{
A batchtools TORQUE backend resolves futures in parallel via a TORQUE/PBS job scheduler
}
\details{
Batchtools torque futures use \pkg{batchtools} cluster functions
created by \code{\link[batchtools:makeClusterFunctionsTORQUE]{batchtools::makeClusterFunctionsTORQUE()}}, which requires
that TORQUE commands \code{qsub}, \code{qselect}, and \code{qdel} are installed on
the current machine.

The default template script \code{templates/torque.tmpl} can be found in:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("templates", "torque.tmpl", package = "future.batchtools")
}\if{html}{\out{</div>}}

and comprise:

\if{html}{\out{<div class="sourceCode bash">}}\preformatted{#!/bin/bash
######################################################################
# A batchtools launch script template for TORQUE/PBS
#
# Author: Henrik Bengtsson 
######################################################################

## Job name:
#PBS -N <\%= job.name \%>

## Direct streams to logfile:
#PBS -o <\%= log.file \%>

## Merge standard error and output:
#PBS -j oe

## Email on abort (a) and termination (e), but not when starting (b)
#PBS -m ae

## Resources needed:
<\% if (length(resources) > 0) \{
  opts <- unlist(resources, use.names = TRUE)
  opts <- sprintf("\%s=\%s", names(opts), opts)
  cat(sprintf("#PBS -l \%s\\n", opts))
\} \%>

echo "Batchtools job name: '<\%= job.name \%>'"

echo "Session information:"
date
hostname
which Rscript
Rscript --version
Rscript -e ".libPaths()"

## Launch R and evaluate the batchtools R job
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ..."
Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")'
res=$?
echo " - exit code: $\{res\}"
echo "Command: Rscript -e 'batchtools::doJobCollection("<\%= uri \%>")' ... done"

## End-of-job summary
qstat -f "$\{PBS_JOBID\}"

## Relay the exit code from Rscript
exit "$\{res\}"
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Limit runtime to 3 minutes and memory to 200 MiB per future
plan(future.batchtools::batchtools_torque, resources = list(walltime = "00:03:00", mem = "200mb"))

message("Main process ID: ", Sys.getpid())

f <- future(Sys.getpid())
pid <- value(f)
message("Worker process ID: ", pid)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \url{https://en.wikipedia.org/wiki/TORQUE}
}
}
