% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtoolsFutureError.R
\name{BatchtoolsFutureError}
\alias{BatchtoolsFutureError}
\title{FutureError class for errors related to BatchtoolsFuture:s}
\usage{
BatchtoolsFutureError(...)
}
\arguments{
\item{\ldots}{Arguments passed to \code{\link[future]{FutureError}}}
}
\description{
FutureError class for errors related to BatchtoolsFuture:s
}
\keyword{internal}
