% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools_template.R
\name{BatchtoolsTemplateFutureBackend}
\alias{BatchtoolsTemplateFutureBackend}
\alias{BatchtoolsLsfFutureBackend}
\alias{BatchtoolsOpenLavaFutureBackend}
\alias{BatchtoolsSGEFutureBackend}
\alias{BatchtoolsSlurmFutureBackend}
\alias{BatchtoolsTorqueFutureBackend}
\alias{batchtools_lsf}
\alias{batchtools_openlava}
\alias{batchtools_torque}
\title{Batchtools futures for LSF, OpenLava, SGE, Slurm, TORQUE etc.}
\usage{
BatchtoolsTemplateFutureBackend(
  ...,
  template = NULL,
  type = c("lsf", "openlava", "sge", "slurm", "torque")
)
}
\arguments{
\item{template}{(optional) A batchtools template file or a template string
(in \pkg{brew} format).  If not specified, it is left to the
\pkg{batchtools} package to locate such file using its search rules.}

\item{type}{(character string) Type of job scheduler.}

\item{\ldots}{Additional arguments passed to \code{\link[=BatchtoolsFutureBackend]{BatchtoolsFutureBackend()}}.}
}
\value{
An object of class \code{BatchtoolsFutureBackend}.
}
\description{
Batchtools futures for LSF, OpenLava, SGE, Slurm, TORQUE etc. are
asynchronous multiprocess futures that will be evaluated on a compute
cluster via a job scheduler.
}
\details{
These type of batchtools futures rely on batchtools backends set
up using the following \pkg{batchtools} functions:
\itemize{
\item \code{\link[batchtools:makeClusterFunctionsLSF]{batchtools::makeClusterFunctionsLSF()}} for
\href{https://en.wikipedia.org/wiki/Platform_LSF}{Load Sharing Facility (LSF)}
\item \code{\link[batchtools:makeClusterFunctionsOpenLava]{batchtools::makeClusterFunctionsOpenLava()}} for
\href{https://en.wikipedia.org/wiki/OpenLava}{OpenLava}
\item \code{\link[batchtools:makeClusterFunctionsSGE]{batchtools::makeClusterFunctionsSGE()}} for
\href{https://en.wikipedia.org/wiki/Oracle_Grid_Engine}{Sun/Oracle Grid Engine (SGE)}
\item \code{\link[batchtools:makeClusterFunctionsSlurm]{batchtools::makeClusterFunctionsSlurm()}} for
\href{https://en.wikipedia.org/wiki/Slurm_Workload_Manager}{Slurm}
\item \code{\link[batchtools:makeClusterFunctionsTORQUE]{batchtools::makeClusterFunctionsTORQUE()}} for
\href{https://en.wikipedia.org/wiki/TORQUE}{TORQUE} / PBS
}
}
