% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources_OP.R
\name{\%resources\%}
\alias{\%resources\%}
\title{Temporarily tweaks the resources for the current batchtools strategy}
\usage{
fassignment \%resources\% tweaks
}
\arguments{
\item{fassignment}{The future assignment, e.g.
\code{x \%<-\% \{ expr \}}.}

\item{tweaks}{A named list (or vector) of resource
batchtools parameters that should be changed relative to
the current strategy.}
}
\description{
Temporarily tweaks the resources for the current batchtools strategy
}
\keyword{internal}
